refactor: consolidate OfferRequestBlock form and add smooth navigation
